requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-appconf==1.0.2
django-compressor==2.2
djangorestframework==3.7.7
idna==2.6
mailchimp3==2.1.0
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
urllib3==1.22
